further interface removal of partner repo (WIP)
