+++ src/main/scala/krs/user/InfrastructureModule.scala
@@ -45,6 +45,5 @@ object UserFileRepository {
 
 trait InfrastructureModule { this: DomainModule =>
   val repository = UserMemoryRepository()
-  val partnerRepository = new krs.partner.Injector()
   val eligibilityApi = krs.eligibility.EligibilitySystem
 }
